refactor(web.routing): move metadata objects to Assimalign.Cohesion.Web.Routing.Metadata namespace [L03.01.02.01.06]
Follow-up to the Metadata/ folder move: the metadata carriers now carry a
.Metadata sub-namespace, mirroring the Patterns/Policies areas, instead of
sitting in the root namespace. RouterRouteMetadataCollection,
RouteHostMetadata, and RouteHostConstraint move to
Assimalign.Cohesion.Web.Routing.Metadata; the read contract
IRouterRouteMetadataCollection stays in Abstractions/ at the root namespace
(interface-first — consumers depending only on the contract are untouched).

Public-namespace change for the three concrete types: callers that
construct them add 'using Assimalign.Cohesion.Web.Routing.Metadata;'
(updated in-repo: Route, Router, RouteMatchFeature, the routing extensions,
the two abstraction doc-crefs, six test files, and the Cookie auth test
double). Sibling PRs #883/#885 that construct RouterRouteMetadataCollection
will pick up the same one-line using on rebase.
