+++ resources/Web/Assimalign.Cohesion.Web.Routing/src/Abstractions/IRouterRoute.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 
 using Assimalign.Cohesion.Http;
+using Assimalign.Cohesion.Web.Routing.Metadata;
 
 namespace Assimalign.Cohesion.Web.Routing;
