+++ resources/Web/Assimalign.Cohesion.Web.Routing/tests/RouteHostConstraintTests.cs
@@ -1,5 +1,6 @@
 using Assimalign.Cohesion.Http;
 using Assimalign.Cohesion.Web.Routing.Exceptions;
+using Assimalign.Cohesion.Web.Routing.Metadata;
 using Shouldly;
 using Xunit;
