+++ resources/Web/Assimalign.Cohesion.Web.Routing/tests/RouteHostMatchingTests.cs
@@ -2,6 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using Assimalign.Cohesion.Http;
+using Assimalign.Cohesion.Web.Routing.Metadata;
 using Assimalign.Cohesion.Web.Routing.Tests.TestObjects;
 using Shouldly;
 using Xunit;
