+++ resources/Web/Assimalign.Cohesion.Web.Routing/tests/RouteMatchFeatureTests.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 
 using Assimalign.Cohesion.Http;
+using Assimalign.Cohesion.Web.Routing.Metadata;
 using Assimalign.Cohesion.Web.Routing.Tests.TestObjects;
 
 using Shouldly;
