+++ resources/Web/Assimalign.Cohesion.Web.Routing/tests/UseRoutingMiddlewareTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading.Tasks;
 
 using Assimalign.Cohesion.Http;
+using Assimalign.Cohesion.Web.Routing.Metadata;
 using Assimalign.Cohesion.Web.Routing.Tests.TestObjects;
 
 using Shouldly;
